Fix race condition in parallel tests.
Parallel executor tests checks that when timeout occurds parallel
executor will exit even if long lasting operation scheduled in separate
go routine is running. This go routine might finish after the whole test is
finished and this is expected behaviour. In this go routine we should
not use t.Log which will be invoked at the end of the action because the
test might already exit and this can cause panic.