Change BackendService beta API tests
Fields present in beta.BackendService API might be promoted to v1 any time and
it is hard to follow that changes. Having a test that checks that fields are
present only in beta API and checkst that conversion between Ga and Beta version
fails are unstable and hard to maintain. Removing this logic from tests.